requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==3.7.0
django-decouple==2.1
django-filter==2.4.0
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.3.1
psycopg2==2.9.1
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.3.0
